# Bring back List filter/filterNot optimization to 2.13