This change was brought to 2.12 from https://github.com/scala/scala/pull/5664/files
but as it broke MIMA check, it was reverted from: https://github.com/scala/scala/commit/76babbb6728e5daf4fb95444273f6f8e0a4098d7