@retronym 2018-03-07
The test failure:

```
[test] !!  615 - run/repl-colon-type.scala                 [output differs]
[test] % diff /home/jenkins/workspace/scala-2.13.x-validate-main@2/test/files/run/repl-colon-type-run.log /home/jenkins/workspace/scala-2.13.x-validate-main@2/test/files/run/repl-colon-type.check
[test] @@ -79,7 +79,7 @@ TypeRef(
[test]      )
[test]      TypeRef(
[test]        TypeSymbol(
[test] -        sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with FilteredTraversableInternal[A,List[A]] with Serializable
[test] +        sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with Serializable
[test]  
[test]        )
[test]        args = List(
[test] @@ -146,7 +146,7 @@ TypeRef(
[test]        args = List(
[test]          TypeRef(
[test]            TypeSymbol(
[test] -            sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with FilteredTraversableInternal[A,List[A]] with Serializable
[test] +            sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with Serializable
[test]  
[test]            )
[test]            args = List(
[test] @@ -179,7 +179,7 @@ PolyType(
[test]            args = List(
[test]              TypeRef(
[test]                TypeSymbol(
[test] -                sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with FilteredTraversableInternal[A,List[A]] with Serializable
[test] +                sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with Serializable
[test]  
[test]                )
[test]                args = List(TypeParamTypeRef(TypeParam(T <: AnyVal)))
[test] @@ -202,7 +202,7 @@ PolyType(
[test]      params = List(TermSymbol(x: T), TermSymbol(y: List[U]))
[test]      resultType = TypeRef(
[test]        TypeSymbol(
[test] -        sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with FilteredTraversableInternal[A,List[A]] with Serializable
[test] +        sealed abstract class List[+A] extends AbstractSeq[A] with LinearSeq[A] with Product with GenericTraversableTemplate[A,List] with LinearSeqOptimized[A,List[A]] with Serializable
[test]  
[test]        )
[test]        args = List(TypeParamTypeRef(TypeParam(U >: T)))
[test] 
...
```

Could be addressed by updating the checkfile `sbt "partest --update-check run/repl-colon-type.scala`

But I question why you're introducing `FilteredTraversableInternal` as part of this change? Is it still relevant in the 2.13.x branch where we don't have binary compatiblity constraints? I'd suggest just adding the new implementations to `List` directly.