@retronym 2018-03-07
Minimizes to:

```
object TreeSetTest {
  def test(subject: scala.collection.immutable.TreeSet[Int]) = subject.zip(subject)
}
```

```
⚡ qscalac  sandbox/test.scala
sandbox/test.scala:2: error: ambiguous implicit values:
 both method newCanBuildFrom in object SortedSet of type [A](implicit ord: Ordering[A])scala.collection.generic.CanBuildFrom[scala.collection.immutable.SortedSet.Coll,A,scala.collection.immutable.SortedSet[A]]
 and method newCanBuildFrom in class SortedSetFactory of type [A](implicit ord: Ordering[A])scala.collection.generic.CanBuildFrom[scala.collection.immutable.TreeSet.Coll,A,scala.collection.immutable.TreeSet[A]]
 match expected type scala.collection.generic.CanBuildFrom[scala.collection.immutable.TreeSet[Int],(Int, Int),That]
  def test(subject: scala.collection.immutable.TreeSet[Int]) = subject.zip(subject)
                                                                          ^
one error found
```

I haven't minimized whatever it is that makes this sometimes pass and sometimes fail.