@ackratos 2018-03-07
@retronym okay, I will do similar thing as https://github.com/scala/scala/pull/6372