@ackratos 2018-03-09
performance:

baseline 2.13.x 90ffe626ecd7b888794740f4b4b143d4728d8232
```
[info] # Run complete. Total time: 00:31:25
[info] 
[info] Benchmark                            (size)  Mode  Cnt      Score      Error  Units
[info] ListBenchmark.filter_exc_last             0  avgt   20     39.964 ±   11.377  ns/op
[info] ListBenchmark.filter_exc_last             1  avgt   20    108.877 ±   40.773  ns/op
[info] ListBenchmark.filter_exc_last            10  avgt   20    419.570 ±   79.296  ns/op
[info] ListBenchmark.filter_exc_last           100  avgt   20   2711.160 ± 1578.897  ns/op
[info] ListBenchmark.filter_exc_last          1000  avgt   20  17745.196 ± 3647.482  ns/op
[info] ListBenchmark.filter_exc_mid              0  avgt   20     22.168 ±    9.908  ns/op
[info] ListBenchmark.filter_exc_mid              1  avgt   20     25.270 ±    5.292  ns/op
[info] ListBenchmark.filter_exc_mid             10  avgt   20    139.599 ±    6.421  ns/op
[info] ListBenchmark.filter_exc_mid            100  avgt   20   1321.888 ±   30.671  ns/op
[info] ListBenchmark.filter_exc_mid           1000  avgt   20  12975.399 ±  527.259  ns/op
[info] ListBenchmark.filter_excludeAll           0  avgt   20     15.082 ±    0.868  ns/op
[info] ListBenchmark.filter_excludeAll           1  avgt   20     21.121 ±    1.544  ns/op
[info] ListBenchmark.filter_excludeAll          10  avgt   20     66.097 ±    2.047  ns/op
[info] ListBenchmark.filter_excludeAll         100  avgt   20    530.223 ±    5.490  ns/op
[info] ListBenchmark.filter_excludeAll        1000  avgt   20   5158.415 ±  466.000  ns/op
[info] ListBenchmark.filter_from_mid             0  avgt   20     16.576 ±    1.393  ns/op
[info] ListBenchmark.filter_from_mid             1  avgt   20     30.414 ±    2.772  ns/op
[info] ListBenchmark.filter_from_mid            10  avgt   20    119.492 ±    6.626  ns/op
[info] ListBenchmark.filter_from_mid           100  avgt   20    965.398 ±   22.339  ns/op
[info] ListBenchmark.filter_from_mid          1000  avgt   20  10061.495 ±  776.855  ns/op
[info] ListBenchmark.filter_includeAll           0  avgt   20     14.516 ±    0.651  ns/op
[info] ListBenchmark.filter_includeAll           1  avgt   20     27.210 ±    1.612  ns/op
[info] ListBenchmark.filter_includeAll          10  avgt   20    156.706 ±   28.959  ns/op
[info] ListBenchmark.filter_includeAll         100  avgt   20   1456.659 ±  123.793  ns/op
[info] ListBenchmark.filter_includeAll        1000  avgt   20  12456.409 ± 1112.700  ns/op
[info] ListBenchmark.filter_only_last            0  avgt   20     16.045 ±    0.169  ns/op
[info] ListBenchmark.filter_only_last            1  avgt   20     28.803 ±    0.403  ns/op
[info] ListBenchmark.filter_only_last           10  avgt   20     81.480 ±    0.802  ns/op
[info] ListBenchmark.filter_only_last          100  avgt   20    608.914 ±   13.134  ns/op
[info] ListBenchmark.filter_only_last         1000  avgt   20   6126.541 ±  104.027  ns/op
```

This PR Optimization
```
[info] # Run complete. Total time: 00:31:27
[info] 
[info] Benchmark                            (size)  Mode  Cnt      Score      Error  Units
[info] ListBenchmark.filter_exc_last             0  avgt   20      4.554 ±    1.319  ns/op
[info] ListBenchmark.filter_exc_last             1  avgt   20      6.201 ±    0.228  ns/op
[info] ListBenchmark.filter_exc_last            10  avgt   20    189.247 ±   10.173  ns/op
[info] ListBenchmark.filter_exc_last           100  avgt   20   1847.792 ±   40.684  ns/op
[info] ListBenchmark.filter_exc_last          1000  avgt   20  14454.780 ± 6192.910  ns/op
[info] ListBenchmark.filter_exc_mid              0  avgt   20      4.751 ±    1.502  ns/op
[info] ListBenchmark.filter_exc_mid              1  avgt   20      6.071 ±    0.113  ns/op
[info] ListBenchmark.filter_exc_mid             10  avgt   20     62.440 ±   12.901  ns/op
[info] ListBenchmark.filter_exc_mid            100  avgt   20   1070.520 ±   11.481  ns/op
[info] ListBenchmark.filter_exc_mid           1000  avgt   20   5155.931 ±  212.738  ns/op
[info] ListBenchmark.filter_excludeAll           0  avgt   20      3.860 ±    0.118  ns/op
[info] ListBenchmark.filter_excludeAll           1  avgt   20      6.280 ±    0.714  ns/op
[info] ListBenchmark.filter_excludeAll          10  avgt   20     20.656 ±    1.195  ns/op
[info] ListBenchmark.filter_excludeAll         100  avgt   20    223.719 ±    3.060  ns/op
[info] ListBenchmark.filter_excludeAll        1000  avgt   20   2663.433 ±   74.063  ns/op
[info] ListBenchmark.filter_from_mid             0  avgt   20      3.673 ±    0.114  ns/op
[info] ListBenchmark.filter_from_mid             1  avgt   20      5.027 ±    0.094  ns/op
[info] ListBenchmark.filter_from_mid            10  avgt   20    119.516 ±   15.875  ns/op
[info] ListBenchmark.filter_from_mid           100  avgt   20   1133.278 ±   85.667  ns/op
[info] ListBenchmark.filter_from_mid          1000  avgt   20  10803.648 ±  256.929  ns/op
[info] ListBenchmark.filter_includeAll           0  avgt   20      4.131 ±    0.345  ns/op
[info] ListBenchmark.filter_includeAll           1  avgt   20      5.277 ±    0.771  ns/op
[info] ListBenchmark.filter_includeAll          10  avgt   20     22.902 ±    3.330  ns/op
[info] ListBenchmark.filter_includeAll         100  avgt   20    261.138 ±    4.574  ns/op
[info] ListBenchmark.filter_includeAll        1000  avgt   20   2669.859 ±   99.142  ns/op
[info] ListBenchmark.filter_only_last            0  avgt   20      4.041 ±    0.258  ns/op
[info] ListBenchmark.filter_only_last            1  avgt   20      5.777 ±    1.229  ns/op
[info] ListBenchmark.filter_only_last           10  avgt   20     29.416 ±    1.619  ns/op
[info] ListBenchmark.filter_only_last          100  avgt   20    277.224 ±   27.001  ns/op
[info] ListBenchmark.filter_only_last         1000  avgt   20   2605.717 ±   65.524  ns/op
```